kiwibike/kiwibike/requirements.txt
amqp==2.4.0
appnope==0.1.0
autopep8==1.4.3
backcall==0.1.0
billiard==3.5.0.5
boto==2.49.0
boto3==1.9.86
botocore==1.12.86
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
click==7.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.2
Django==2.1.5
django-cors-headers==2.4.0
django-crontab==0.7.1
django-s3-storage==0.12.4
django-filter==2.1.0
django-redis==4.10.0
django-modeltranslation==0.13b1
django-queryinspect==1.0.0
django-rest-framework==0.1.0
django-ses==0.8.9
django-silk==3.0.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
drf-yasg==1.13.0
flake8==3.7.3
flex==6.13.2
future==0.17.1
gprof2dot==2017.9.19
idna==2.8
inflection==0.3.1
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jsondiff==1.1.2
jsonpickle==1.1
jsonpointer==2.0
kombu==4.2.2.post1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pip-autoremove==0.9.1
prompt-toolkit==2.0.8
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.0
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
raven==6.10.0
redis==3.1.0
requests==2.21.0
rest-condition==1.0.3
rfc3987==1.3.8
ruamel.yaml==0.15.87
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.2.4
strict-rfc3339==0.7
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
vine==1.2.0
wcwidth==0.1.7
rollbar==0.14.6
